requirements.txt
tqdm
datetime
requests==2.27.0
quart_cors
uvicorn
httptools
pyyaml
quart
hypercorn==0.14.3
python-dotenv
apscheduler
firebase-admin
openai
newsapi-python
uuid